+++ test/unit_tests/test_EditorCellInvariants.cpp
@@ -52,7 +52,7 @@
 #include <wx/log.h>
 
 #include "Configuration.h"
-#include "Worksheet.h"
+#include "worksheet/Worksheet.h"
 #include "cells/CellList.h"
 #include "cells/GroupCell.h"
 #include "cells/EditorCell.h"