+++ test/unit_tests/test_WorksheetFind.cpp
@@ -37,7 +37,7 @@
 #include <wx/log.h>
 
 #include "Configuration.h"
-#include "Worksheet.h"
+#include "worksheet/Worksheet.h"
 #include "cells/EditorCell.h"
 #include "cells/GroupCell.h"
 #include "cells/TextCell.h"